python_backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.1
